+++ lib/htmlassembler.spec.js
@@ -1,7 +1,7 @@
 import { HTMLAssembler } from './htmlassembler.js';
 import chai from 'chai';
 
-import { htmlEqualMarkup } from './util/domequalmarkup.js';
+import { htmlEqualMarkup } from './util/htmlequalmarkup.js';
 
 const { assert } = chai;
 const {
@@ -21,11 +21,68 @@ const {
 
 const assembler = new HTMLAssembler;
 
-describe('HTML assembler', () => {
+describe('HTMLDOM assembler', () => {
 
     describe('createElement', () => {
 
         describe('built-in global attributes', () => {
+
+            describe('id', () => {
+                const id = 'element_0';
+                const element = assembler.createElement('span', { id });
+
+                it('proper number of attributes', () => {
+                    assert(element.hasAttributes(), 'element has attributes');
+                    assert.equal(element.attributes.length, 1);
+                });
+                it('proper `id` attribute', () => {
+                    assert.equal(element.id, id);
+                });
+                it('document finds it', () => {
+                    document.body.appendChild(element);
+                    assert.equal(document.getElementById(id), element);
+                    document.body.removeChild(element);
+                });
+            });
+
+            describe('className', () => {
+                const className = 'foo bar wiz';
+                const element = assembler.createElement('element', { className });
+                it('proper number of attributes', () => {
+                    assert(element.hasAttributes(), 'element has attributes');
+                    assert.equal(element.attributes.length, 1);
+                });
+                it('proper `className` attribute', () => {
+                    assert.equal(element.className, className);
+                    assert.equal(element.getAttribute('class'), className);
+                });
+                it('classList interface works properly', () => {
+                    assert(element.classList.contains('foo'), 'contains class `foo`');
+                    assert(element.classList.contains('bar'), 'contains class `bar`');
+                    assert(element.classList.contains('wiz'), 'contains class `wiz`');
+                });
+                it('document finds it', () => {
+                    document.body.appendChild(element);
+                    assert.equal(document.getElementsByClassName('foo')[0], element);
+                    assert.equal(document.getElementsByClassName('bar')[0], element);
+                    assert.equal(document.getElementsByClassName('wiz')[0], element);
+                    document.body.removeChild(element);
+                });
+            });
+
+            describe('textContent', () => {
+                const init = { textContent : 'element textContent' };
+                const element = assembler.createElement('element', init);
+                it('proper number of child nodes', () => {
+                    assert(element.hasChildNodes(), 'element has child nodes');
+                    assert.equal(element.childNodes.length, 1);
+                });
+                it('proper children', () => {
+                    assert.equal(element.firstChild.constructor, Text);
+                    assert.equal(element.firstChild.textContent, 'element textContent');
+                    assert.equal(element.textContent, 'element textContent');
+                });
+            });
 
             describe('accessKey', () => {
                 const init = { accessKey : 'A' };
@@ -310,6 +367,58 @@ describe('HTML assembler', () => {
         });
 
         describe('adapted interface', () => {
+
+            describe('attrset', () => {
+                const init = { attrset : { selected : '' } };
+                const element = assembler.createElement('option', init);
+                it('proper number of attributes', () => {
+                    assert(element.hasAttributes(), 'element has attributes');
+                    assert.equal(element.attributes.length, 1);
+                });
+                it('proper attributes assigned', () => {
+                    assert.equal(element.getAttribute('selected'), '');
+                });
+            });
+
+            describe('children', () => {
+                const child = new HTMLAssembler;
+                child.createElement('span');
+                const element = assembler.createElement('div', {
+                    children : [
+                        0,
+                        assembler.createElement('button'),
+                        '',
+                        'text node as string',
+                        NaN,
+                        document.createTextNode('created text node'),
+                        null,
+                        document.createElement('hr'),
+                        false,
+                        document.createComment('Simple DOM Comment node'),
+                        undefined,
+                        child
+                    ]
+                });
+                const childNodes = element.childNodes;
+                it('proper number of child nodes', () => {
+                    assert(element.hasChildNodes(), 'element has child nodes');
+                    assert.equal(childNodes.length, 6);
+                });
+                it('Proper child nodes', () => {
+                    assert.equal(childNodes[0].constructor, HTMLButtonElement);
+                    assert.equal(childNodes[0].tagName, 'BUTTON');
+                    assert.equal(childNodes[1].constructor, Text);
+                    assert.equal(childNodes[1].textContent, 'text node as string');
+                    assert.equal(childNodes[2].constructor, Text);
+                    assert.equal(childNodes[2].nodeValue, 'created text node');
+                    assert.equal(childNodes[3].constructor, HTMLHRElement);
+                    assert.equal(childNodes[3].tagName, 'HR');
+                    assert.equal(childNodes[4].constructor, Comment);
+                    assert.equal(childNodes[4].textContent, 'Simple DOM Comment node');
+                    assert.equal(childNodes[5].constructor, HTMLSpanElement);
+                    assert.equal(childNodes[5].tagName, 'SPAN');
+                });
+            });
 
             describe('dataset', () => {
                 const init = { dataset : { camelCased : 'custom attribute' } };
