+++ lib/htmldom.spec.js
@@ -9,7 +9,7 @@ import {
     htmldom, span
 } from './htmldom';
 
-import { htmlEqualMarkup } from './util/domequalmarkup';
+import { htmlEqualMarkup } from './util/htmlequalmarkup';
 
 import chai from 'chai';
