+++ lib/index.js
@@ -2,7 +2,5 @@
  * @module htmlmodule
  */
 export * from './nodeinit';
-export * from './domassembler';
 export * from './htmlassembler';
-export * from './xmldom';
 export * from './htmldom';
