+++ lib/index.spec.js
@@ -1,5 +1,3 @@
 import './nodeinit.spec';
-import './domassembler.spec';
 import './htmlassembler.spec';
-import './xmldom.spec';
 import './htmldom.spec';
